Infrastructure/CliniGraph/requirements.txt
agent==0.1.3
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
backoff==2.2.1
blis==1.3.3
CacheControl==0.14.4
catalogue==2.0.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cloudpathlib==0.23.0
confection==1.3.3
cryptography==46.0.7
cymem==2.0.13
dataclasses-json==0.6.7
deepeval==3.9.7
distro==1.9.0
docopt==0.6.2
execnet==2.1.2
fastapi==0.136.0
filelock==3.29.0
filetype==1.2.0
firebase_admin==7.4.0
frozenlist==1.8.0
fsspec==2026.3.0
google-api-core==2.30.3
google-auth==2.49.2
google-cloud-core==2.5.1
google-cloud-firestore==2.27.0
google-cloud-storage==3.10.1
google-crc32c==1.8.0
google-genai==1.73.1
google-resumable-media==2.8.2
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.80.0
h11==0.16.0
h2==4.3.0
hf-xet==1.4.3
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.11.0
hyperframe==6.1.0
idna==3.12
importlib_metadata==8.7.1
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.14.0
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain==1.2.15
langchain-classic==1.0.4
langchain-community==0.4.1
langchain-core==1.3.0
langchain-google-genai==4.2.2
langchain-mcp-adapters==0.2.2
langchain-openai==1.1.16
langchain-pinecone==0.2.13
langchain-text-splitters==1.1.2
langgraph==1.1.9
langgraph-checkpoint==4.0.2
langgraph-prebuilt==1.0.10
langgraph-sdk==0.3.13
langsmith==0.7.33
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mcp==1.27.0
mdurl==0.1.2
msgpack==1.1.2
multidict==6.7.1
murmurhash==1.0.15
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.4.4
openai==2.32.0
opentelemetry-api==1.41.0
opentelemetry-sdk==1.41.0
opentelemetry-semantic-conventions==0.62b0
orjson==3.11.8
ormsgpack==1.12.2
packaging==24.2
phonenumbers==9.0.28
pinecone==7.3.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
pipreqs==0.4.13
pluggy==1.6.0
portalocker==3.2.0
posthog==7.13.0
preshed==3.0.13
presidio_analyzer==2.2.362
presidio_anonymizer==2.2.362
propcache==0.4.1
proto-plus==1.27.2
protobuf==6.33.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic-settings==2.14.0
pydantic_core==2.46.3
pyfiglet==1.0.4
Pygments==2.20.0
PyJWT==2.12.1
pypdf==6.10.2
pytest==9.0.3
pytest-asyncio==1.3.0
pytest-repeat==0.9.4
pytest-rerunfailures==16.1
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.26
PyYAML==6.0.3
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
requests-file==3.0.1
requests-toolbelt==1.0.0
rich==14.3.4
rpds-py==0.30.0
safetensors==0.7.0
sentry-sdk==2.58.0
setuptools==82.0.1
shellingham==1.5.4
simsimd==6.5.16
six==1.17.0
smart_open==7.6.0
sniffio==1.3.1
spacy==3.8.14
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.49
srsly==2.5.3
sse-starlette==3.3.4
starlette==1.0.0
tabulate==0.9.0
tenacity==9.1.4
thinc==8.3.13
tiktoken==0.12.0
tldextract==5.3.1
tokenizers==0.22.2
tqdm==4.67.3
transformers==5.5.4
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.45.0
wasabi==1.1.3
weasel==1.0.0
websockets==16.0
wheel==0.46.3
wrapt==2.1.2
xxhash==3.6.0
yarg==0.1.10
yarl==1.23.0
zipp==3.23.1
zstandard==0.25.0
torch
pymupdf

